+++ crates/store/re_types/tests/types/points3d.rs
@@ -1,55 +1,53 @@
 use std::collections::HashMap;
 
-use re_types::{
-    archetypes::Arrows3D,
-    components::{ClassId, Color, Position3D, Radius, Vector3D},
-    datatypes::Vec3D,
-    Archetype as _, AsComponents as _,
-};
+use re_types::{archetypes::Points3D, components, Archetype as _, AsComponents as _};
+
+use crate::util;
 
 #[test]
 fn roundtrip() {
-    let expected = Arrows3D {
-        vectors: vec![
-            Vector3D(Vec3D([1.0, 2.0, 3.0])),
-            Vector3D(Vec3D([10.0, 20.0, 30.0])),
+    let expected = Points3D {
+        positions: vec![
+            components::Position3D::new(1.0, 2.0, 3.0), //
+            components::Position3D::new(4.0, 5.0, 6.0),
         ],
-        origins: Some(vec![
-            Position3D(Vec3D([4.0, 5.0, 6.0])),    //
-            Position3D(Vec3D([40.0, 50.0, 60.0])), //
-        ]),
         radii: Some(vec![
-            Radius::from(1.0), //
-            Radius::from(10.0),
+            components::Radius::from(42.0), //
+            components::Radius::from(43.0),
         ]),
         colors: Some(vec![
-            Color::from_unmultiplied_rgba(0xAA, 0x00, 0x00, 0xCC), //
-            Color::from_unmultiplied_rgba(0x00, 0xBB, 0x00, 0xDD),
+            components::Color::from_unmultiplied_rgba(0xAA, 0x00, 0x00, 0xCC), //
+            components::Color::from_unmultiplied_rgba(0x00, 0xBB, 0x00, 0xDD),
         ]),
         labels: Some(vec![
             "hello".into(),  //
             "friend".into(), //
         ]),
         class_ids: Some(vec![
-            ClassId::from(126), //
-            ClassId::from(127), //
+            components::ClassId::from(126), //
+            components::ClassId::from(127), //
+        ]),
+        keypoint_ids: Some(vec![
+            components::KeypointId::from(2), //
+            components::KeypointId::from(3), //
         ]),
     };
 
-    let arch = Arrows3D::from_vectors([[1.0, 2.0, 3.0], [10.0, 20.0, 30.0]])
-        .with_origins([[4.0, 5.0, 6.0], [40.0, 50.0, 60.0]])
-        .with_radii([1.0, 10.0])
+    let arch = Points3D::new([(1.0, 2.0, 3.0), (4.0, 5.0, 6.0)])
+        .with_radii([42.0, 43.0])
         .with_colors([0xAA0000CC, 0x00BB00DD])
         .with_labels(["hello", "friend"])
-        .with_class_ids([126, 127]);
+        .with_class_ids([126, 127])
+        .with_keypoint_ids([2, 3]);
     similar_asserts::assert_eq!(expected, arch);
 
     let expected_extensions: HashMap<_, _> = [
-        ("arrows", vec!["rerun.components.Arrow3D"]),
+        ("points", vec!["rerun.components.Position3D"]),
         ("radii", vec!["rerun.components.Radius"]),
         ("colors", vec!["rerun.components.Color"]),
         ("labels", vec!["rerun.components.Text"]),
         ("class_ids", vec!["rerun.components.ClassId"]),
+        ("keypoint_ids", vec!["rerun.components.KeypointId"]),
     ]
     .into();
 
@@ -71,8 +69,6 @@ fn roundtrip() {
         }
     }
 
-    let deserialized = Arrows3D::from_arrow(serialized).unwrap();
+    let deserialized = Points3D::from_arrow(serialized).unwrap();
     similar_asserts::assert_eq!(expected, deserialized);
 }
-
-mod util;
