+++ crates/store/re_types/tests/types/util.rs
@@ -1,12 +1,10 @@
-#[allow(dead_code)]
-pub fn assert_extensions(array: &dyn ::arrow2::array::Array, expected: &[&str]) {
+pub(crate) fn assert_extensions(array: &dyn ::arrow2::array::Array, expected: &[&str]) {
     let mut extracted = vec![];
     extract_extensions(array.data_type(), &mut extracted);
     similar_asserts::assert_eq!(expected, extracted);
 }
 
-#[allow(dead_code)]
-pub fn extract_extensions(datatype: &::arrow2::datatypes::DataType, acc: &mut Vec<String>) {
+pub(crate) fn extract_extensions(datatype: &::arrow2::datatypes::DataType, acc: &mut Vec<String>) {
     match datatype {
         arrow2::datatypes::DataType::List(field)
         | arrow2::datatypes::DataType::FixedSizeList(field, _)
