+++ crates/store/re_types/tests/types/validity.rs
